Use ActionButton for firefox addon, Cleanup, Fix build script

+++ build.sh
@@ -8,18 +8,17 @@ cd ../..
 #Собираем XPI для firefox desktop
 cd firefox
 cfx xpi
-mv firefox.xpi ../bin/chas-correct-firefox-desktop.xpi
+mv chas-correct.xpi ../bin/chas-correct-firefox-desktop.xpi
 #И для mobile
 cfx xpi --force-mobile
-mv firefox.xpi ../bin/chas-correct-firefox-mobile.xpi
+mv chas-correct.xpi ../bin/chas-correct-firefox-mobile.xpi
 cd ..
 
 #Для хрома
-./crxmake.sh chrome ../correct.pem
-mv chrome.crx bin/chas-correct-chrome-desktop.crx
-cp bin/chas-correct-chrome-desktop.crx bin/chas-correct.zip
-
-
-
-
-
+if [ -f chrome.pem ]; then
+  ./crxmake.sh chrome chrome.pem
+  mv chrome.crx bin/chas-correct-chrome-desktop.crx
+  cp bin/chas-correct-chrome-desktop.crx bin/chas-correct.zip
+else
+  echo '"chrome.pem" file not found. Skipping Chrome ext build.'
+fi
